Delete the comments, move the bookForm to containers
